pyproject.toml
[tool.poetry]
name = "mem-translator"
version = "0.1.0"
description = "A meme translation service"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.13.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-telegram-bot = "21.10"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
pytesseract = "^0.3.10"
deep-translator = "^1.11.4"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
boto3 = "^1.29.3"
pydantic = "^2.5.1"
loguru = "^0.7.2"
alembic = "^1.13.1"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
